Fix checkstyle and PMD errors

+++ app/src/main/java/com/nervousfish/nervousfish/activities/MainActivity.java
@@ -7,7 +7,6 @@ import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -23,7 +22,6 @@ import com.nervousfish.nervousfish.modules.database.IDatabase;
 import com.nervousfish.nervousfish.service_locator.IServiceLocator;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -71,7 +69,7 @@ public final class MainActivity extends AppCompatActivity {
 
         try {
             lv.setAdapter(new ContactListAdapter(this, serviceLocator.getDatabase().getAllContacts()));
-        } catch (IOException e) {
+        } catch (final IOException e) {
             e.printStackTrace();
         }
 
@@ -81,44 +79,58 @@ public final class MainActivity extends AppCompatActivity {
      * Temporarily fill the database with demo data for development.
      */
     private void fillDatabaseWithDemoData() {
-        IDatabase database = serviceLocator.getDatabase();
+        final IDatabase database = serviceLocator.getDatabase();
         try {
-            Contact c = new Contact("Eric", new SimpleKey("jdfs09jdfs09jfs0djfds9jfsd0"));
-            database.deleteContact(c);
+
+            final Contact a = new Contact("Eric", new SimpleKey("jdfs09jdfs09jfs0djfds9jfsd0"));
+            final Contact b = new Contact("Stas", new SimpleKey("4ji395j495i34j5934ij534i"));
+            final Contact c = new Contact("Joost", new SimpleKey("dnfh4nl4jknlkjnr4j34klnk3j4nl"));
+            final Contact d = new Contact("Kilian", new SimpleKey("sdjnefiniwfnfejewjnwnkenfk32"));
+            final Contact e = new Contact("Cornel", new SimpleKey("nr23uinr3uin2o3uin23oi4un234ijn"));
+            if(!database.getAllContacts().isEmpty()) {
+                database.deleteContact(a);
+                database.deleteContact(b);
+                database.deleteContact(c);
+                database.deleteContact(d);
+                database.deleteContact(e);
+            }
+            database.addContact(a);
+            database.addContact(b);
             database.addContact(c);
-            c = new Contact("Stas", new SimpleKey("4ji395j495i34j5934ij534i"));
-            database.deleteContact(c);
-            database.addContact(c);
-            c = new Contact("Joost", new SimpleKey("dnfh4nl4jknlkjnr4j34klnk3j4nl"));
-            database.deleteContact(c);
-            database.addContact(c);
-            c = new Contact("Kilian", new SimpleKey("sdjnefiniwfnfejewjnwnkenfk32"));
-            database.deleteContact(c);
-            database.addContact(c);
-            c = new Contact("Cornel", new SimpleKey("nr23uinr3uin2o3uin23oi4un234ijn"));
-            database.deleteContact(c);
-            database.addContact(c);
-        } catch (IOException e) {
+            database.addContact(d);
+            database.addContact(e);
+        } catch (final IOException e) {
             e.printStackTrace();
         }
     }
 }
 
-
+/**
+ * An Adapter which converts a list with contacts into List entries.
+ */
 class ContactListAdapter extends ArrayAdapter<Contact> {
 
-    public ContactListAdapter(final Context context, final List<Contact> items) {
-        super(context, 0, items);
+    private final static int MAX_SNIPPET_SIZE = 30;
+    /**
+     * Create and initialize a ContactListAdapter.
+     *
+     * @param context the Context where the ListView is created
+     * @param contacts the list with contacts
+     */
+    ContactListAdapter(final Context context, final List<Contact> contacts) {
+        super(context, 0, contacts);
     }
 
+    /**
+     * {@inheritDoc}
+     */
     @Override
     public View getView(final int position, final View convertView, final ViewGroup parent) {
 
         View v = convertView;
 
         if (v == null) {
-            final LayoutInflater vi;
-            vi = LayoutInflater.from(getContext());
+            final LayoutInflater vi = LayoutInflater.from(getContext());
             v = vi.inflate(R.layout.contact_list_entry, null);
         }
 
@@ -134,7 +146,7 @@ class ContactListAdapter extends ArrayAdapter<Contact> {
 
             if (pubKey != null) {
                 final String publicKey = contact.getPublicKey().getKey();
-                if(publicKey.length() > 30) {
+                if(publicKey.length() > MAX_SNIPPET_SIZE) {
                     final String pubKeySnippet = contact.getPublicKey().getKey().substring(0, 30) + "...";
                     pubKey.setText(pubKeySnippet);
                 } else {
